backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Car renting app"
authors = ["Ayodimeji <ayodimejia1@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
django-debug-toolbar = "^4.4.6"
djangorestframework = "^3.15.2"
python-decouple = "^3.8"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
